pyproject.toml
[tool.poetry]
name = "mafia-bot"
version = "0.1.0"
description = ""
authors = ["Will <draknodred@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11, <3.13"
hikari-lightbulb = "^2.3.5"
pg8000 = "^1.31.2"
networkx = "^3.3"
matplotlib = "^3.9.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
